remember h file

+++ helper.h
@@ -1,2 +1,2 @@
-double plot_map(double outStart, double outEnd, double inStart, double inEnd, double point);
-double map(double outStart, double outEnd, double inStart, double inEnd, double point);
+long double plot_map(long double outStart, long double outEnd, long double inStart, long double inEnd, long double point);
+long double map(long double outStart, long double outEnd, long double inStart, long double inEnd, long double point);
